// samples/web-csr/dressca-frontend/consumer/src/system-common/generated/api-client/models/put-basket-items-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Dressca
 * ECサイトDressca
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface PutBasketItemsRequest {
    'displayItemId': string;
    'quantity': number;
}